Returns the next largest permutation of a given integer
For each digit in the number, loop through the digits before it till we
find a digit less than it.
If we find one such digit, we can return the next largest permutation by
swapping the two, and sorting
the digits from the ones digit upto, but not including the leftmost
digit swapped, in ascending order, left to right.
Otherwise, return that it is not possible.